+++ test/integration/api/websockets_sanity.js
@@ -1552,6 +1552,46 @@ describe(
       );
     });
 
+    it('subscribes with a count - we ensure the event only gets kicked off for the correct amounts', function(callback) {
+      var hits = 0;
+      listenerclient.on(
+        '/2_websockets_embedded_sanity/' + test_id + '/count/2/*',
+        {
+          event_type: 'set',
+          count: 2
+        },
+        function() {
+          hits++;
+        },
+        function(e) {
+          if (e) return callback(e);
+
+          publisherclient.set('/2_websockets_embedded_sanity/' + test_id + '/count/2/1', {
+            property1: 'property1',
+            property2: 'property2',
+            property3: 'property3'
+          });
+
+          publisherclient.set('/2_websockets_embedded_sanity/' + test_id + '/count/2/2', {
+            property1: 'property1',
+            property2: 'property2',
+            property3: 'property3'
+          });
+
+          publisherclient.set('/2_websockets_embedded_sanity/' + test_id + '/count/2/2', {
+            property1: 'property1',
+            property2: 'property2',
+            property3: 'property3'
+          });
+
+          setTimeout(function() {
+            if (hits !== 2) return callback(new Error('hits were over the agreed on 2'));
+            callback();
+          }, 1500);
+        }
+      );
+    });
+
     it('subscribes with a count - we ensure the event only gets kicked off for the correct amounts - negative test', function(callback) {
       var hits = 0;
       //first listen for the change