+++ test/integration/session/session_events.js
@@ -20,16 +20,56 @@ describe(tests.testName(__filename, 3), function() {
     var serviceInstance;
     var stopped = false;
 
+    var checkEventEmitterStructure = function(eventData, protocol, username) {
+      tests.expect(eventData.id).to.be(eventEmitterClientSessionId);
+      tests.expect(eventData.legacyPing).to.be(false);
+      tests.expect(Number.isInteger(eventData.msgCount)).to.be(true);
+      tests.expect(eventData.isEncrypted).to.be(false);
+      tests.expect(eventData.protocol).to.be(protocol || `happn_${tests.package.protocol}`);
+      tests.expect(eventData.tlsEncrypted).to.be(false);
+      tests.expect(eventData.browser).to.be(false);
+      tests.expect(eventData.intraProc).to.be(true);
+      if (username) tests.expect(eventData.user.username).to.be(username);
+    };
+
+    var checkSocketEventStructure = function(eventData, protocol, username) {
+      tests.expect(eventData.id).to.be(socketClientSessionId);
+      tests.expect(eventData.legacyPing).to.be(false);
+      tests.expect(Number.isInteger(eventData.msgCount)).to.be(true);
+      tests.expect(eventData.isEncrypted).to.be(false);
+      tests.expect(eventData.protocol).to.be(protocol || `happn_${tests.package.protocol}`);
+      tests.expect(eventData.tlsEncrypted).to.be(false);
+      tests.expect(eventData.browser).to.be(false);
+      tests.expect(eventData.intraProc).to.be(false);
+      tests.expect(eventData.sourceAddress).to.be('127.0.0.1');
+      tests.expect(eventData.sourcePort > 0).to.be(true);
+      tests.expect(eventData.upgradeUrl != null).to.be(true);
+      if (username) tests.expect(eventData.user.username).to.be(username);
+    };
+
     var checkAllEventsFired = function(callback) {
       return () => {
-        tests.expect(eventsFired['session-configured-socket'].id).to.be(socketClientSessionId);
-        tests.expect(eventsFired['authentic-socket'].id).to.be(socketClientSessionId);
-        tests.expect(eventsFired['disconnect-socket'].id).to.be(socketClientSessionId);
-        tests
-          .expect(eventsFired['session-configured-eventemitter'].id)
-          .to.be(eventEmitterClientSessionId);
-        tests.expect(eventsFired['authentic-eventemitter'].id).to.be(eventEmitterClientSessionId);
-        tests.expect(eventsFired['disconnect-eventemitter'].id).to.be(eventEmitterClientSessionId);
+        checkSocketEventStructure(eventsFired['connect-socket'], 'happn');
+        checkSocketEventStructure(eventsFired['session-configured-socket']);
+
+        if (serviceConfig.secure)
+          checkSocketEventStructure(eventsFired['authentic-socket'], null, '_ADMIN');
+        else checkSocketEventStructure(eventsFired['authentic-socket']);
+
+        if (serviceConfig.secure)
+          checkSocketEventStructure(eventsFired['disconnect-socket'], null, '_ADMIN');
+        else checkSocketEventStructure(eventsFired['disconnect-socket']);
+
+        checkEventEmitterStructure(eventsFired['connect-eventemitter'], 'happn');
+        checkEventEmitterStructure(eventsFired['session-configured-eventemitter']);
+
+        if (serviceConfig.secure)
+          checkEventEmitterStructure(eventsFired['authentic-eventemitter'], null, '_ADMIN');
+        else checkEventEmitterStructure(eventsFired['authentic-eventemitter']);
+
+        if (serviceConfig.secure)
+          checkEventEmitterStructure(eventsFired['disconnect-eventemitter'], null, '_ADMIN');
+        else checkEventEmitterStructure(eventsFired['disconnect-eventemitter']);
 
         tests.expect(JSON.stringify(eventsFired, null, 2).indexOf('token":')).to.be(-1);
         tests.expect(JSON.stringify(eventsFired, null, 2).indexOf('password')).to.be(-1);
@@ -39,12 +79,7 @@ describe(tests.testName(__filename, 3), function() {
 
         if (stopped) return callback();
         stopped = true;
-        return serviceInstance.stop(
-          {
-            reconnect: false
-          },
-          callback
-        );
+        return serviceInstance.stop(callback);
       };
     };
 
@@ -55,6 +90,11 @@ describe(tests.testName(__filename, 3), function() {
         if (e) return callback(e);
 
         serviceInstance = happnInst;
+
+        serviceInstance.services.session.on('connect', function(data) {
+          if (data.intraProc) return (eventsFired['connect-eventemitter'] = data);
+          eventsFired['connect-socket'] = data;
+        });
 
         serviceInstance.services.session.on('authentic', function(data) {
           if (data.info._local) return (eventsFired['authentic-eventemitter'] = data);