+++ RELEASES.md
@@ -755,7 +755,8 @@
 
 11.4.0 2020-04-24
 -----------------
-  - feature - authentic and disconnect events have more client data #300
+  - feature - new connect event #300
+  - feature - connect, authentic, configure-session, disconnect events have more client data #300
   - fix: safeSessionData no longer includes token
   - protocol feature: client now sends happn version up with configure session
   - neatened up security directory changes event names into constants